Fixed the LayoutFeature for Edges to adjust to the correct text size.
As it turned out the default size for edge labels doesn't really matter. More important is the EdgeLayoutFeature, which is called after the creation and update of Edges. The code to calculate the corret size was already there, but it was guarded by a wrong condition (fqName vs. fqInternalName).
Thanks to Dawid for pointing that out!